GP-46586 Set default creditor to newly-created one
This changes the default creditor to the one created as part of
post-install.sh. CiviSEPA sets up its own test creditor as part
of its setup, and that one isn't compatible with de.systopia.contract.

This is a bit of a hack; we should probably remove the test creditor
and get the creditor ID dynamically.